Main Window Class & Roll Functionality
Moved the main window class over to the main.py file. That way the import of the Dice class can be in main and not be mixed up with only the GUI parts in the gui.py file. Also created a basic roll button and function that uses dice.py to print total rolled for now.