Dice Structure & Dice Add/Sub Button Commands
Changed the structure of the dice class in dice.py to include the name of the dice and the amount. Then declared all of the available dice at the bottom. Imported everything into gui.py and fixed the command attributes for the add  and sub buttons to be used by all the dice types. This solves having to create separate functions for every dice type and just having an add and sub only for all instead. However, still needs to create each individual widget for each dice unfortunately.