Roll Feature
Using the roll button in main.py, a roll function is called within the window class that goes through a list of all the dice options, checks if there's at least one being rolled, calls the roll function within the dice class in dice.py, and shows the total of everything rolled. As of now, addition is hardcoded.